requirements.txt
feedparser==6.0.11
requests==2.31.0
python-dateutil==2.8.2


